backend/requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework-simplejwt==5.3.0
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# File handling
python-docx==1.1.0
PyPDF2==3.0.1
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Excel/CSV
openpyxl==3.1.2
pandas==2.1.3

# Security
cryptography==41.0.7

